feat(llc, ui): add participant sorting, pagination and layout modes. (#338)
* refactor actions.


* add initial support for participant sorting, pin, viewportVisibility.


* add tests for call participant state sorting.


* add actions for sendReaction, setParticipantViewportVisibility, setParticipantPinned.


* fix screenshare track mute/unmute.


* minor improvements.


* handle sendReaction action.


* add dark/light theme support in example.


* fix ui issues, add multiple participant layouts.


* [334] reorganize call actions

* [334] introduce action middlewares

* [334] fix permissions, events, opus-dtx

* [334] rename actions

* temp


* Introduce store.dart


* fix query members middleware.


* merge fixes


* fix participant size change, subscription.


* fix merge conflicts.


* finish implementation.


* fix merge conflicts.


* Update packages/stream_video_flutter/lib/src/call_participants/layout/call_participants_grid_view.dart


* Update packages/stream_video_flutter/lib/src/call_participants/screen_share_call_participants_content.dart


---------

+++ packages/stream_video/lib/src/webrtc/model/stats/rtc_audio_source.dart
@@ -30,7 +30,7 @@ class RtcAudioSource extends RtcMediaSource {
   final double? audioLevel;
   final double? totalAudioEnergy;
   final double? totalSamplesDuration;
-  final int? echoReturnLoss;
+  final double? echoReturnLoss;
   final double? echoReturnLossEnhancement;
 
   String? get codecId => id;
